v1.0a0.dev0 advances
