+++ src/datoso/__main__.py
@@ -1,32 +1,27 @@
 """Main entry point for datoso"""
 import argparse
-import configparser
-import json
 import logging
 import os
-import re
 import sys
-from pathlib import Path
-from venv import logger
 
-from tabulate import tabulate
-
-from datoso import ROOT_FOLDER, __app_name__, __version__
-from datoso.commands.doctor import check_module, check_seed
-from datoso.commands.seed import Seed
+from datoso import __version__
+from datoso.commands.argparser import (
+    add_all_seed_parser,
+    add_config_parser,
+    add_dat_parser,
+    add_deduper_parser,
+    add_doctor_parser,
+    add_import_parser,
+    add_log_parser,
+    add_seed_parser,
+)
 from datoso.configuration import config
-from datoso.configuration.configuration import get_seed_name
 from datoso.configuration.logger import enable_logging, set_verbosity
-from datoso.database.models.datfile import Dat
-from datoso.helpers import Bcolors, FileUtils
-from datoso.helpers.plugins import installed_seeds, seed_description
-from datoso.repositories.dedupe import Dedupe
-from datoso.seeds.rules import Rules
-from datoso.seeds.unknown_seed import detect_seed
+from datoso.helpers import Bcolors
 
 #---------Boilerplate to check python version ----------
-if sys.version_info < (3, 10):
-    print('This is a Python 3 script. Please run it with Python 3.9 or above')
+if sys.version_info < (3, 10):  # noqa: UP036
+    print('This is a Python 3 script. Please run it with Python 3.10 or above')
     sys.exit(1)
 
 
@@ -44,94 +39,14 @@ def parse_args() -> argparse.Namespace:
 
     parser.add_argument('-v', '--version', action='store_true', help='show version')
 
-    parser_log = subparser.add_parser('log', help='Show log')
-    parser_log.set_defaults(func=command_log)
-
-    parser_config = subparser.add_parser('config', help='Show configuration')
-    parser_config.add_argument('-s', '--save', action='store_true', help='Save configuration to .datosorc')
-    parser_config.add_argument('-d', '--directory', default='~', choices=['~', '.'], help='Directory to save .datosorc')
-    parser_config.set_defaults(func=command_config)
-    parser_config.add_argument('-ru', '--rules-update', action='store_true', help='Update system rules from GoogleSheets Url')
-    parser_config.add_argument('-mu', '--mia-update', action='store_true', help='Update MIA from GoogleSheets Url')
-
-    group_save = parser_config.add_mutually_exclusive_group()
-    group_save.add_argument('--set', nargs=2, metavar=('configuration', 'value'), help='Set Configuration Option separated by point with new value e.g. <GENERAL.Overwrite> <false>')
-    group_save.add_argument('--get', metavar=('configuration'), help='Get value of Configuration Option.')
-    parser_config.add_argument('-g','--global', action='store_true', help='When set, saves to global config, else to `.datosorc`')
-
-    parser_doctor = subparser.add_parser('doctor', help='Doctor installed seeds')
-    parser_doctor.add_argument('seed', nargs='?', help='Seed to doctor')
-    parser_doctor.set_defaults(func=command_doctor)
-    parser_doctor.add_argument('-r', '--repair', action='store_true', help='Try to repair seed(s)')
-
-    parser_dat = subparser.add_parser('dat', help='Changes configuration in current dats')
-    parser_dat.add_argument('command', nargs='?', help='Command to execute')
-
-    group_dat = parser_dat.add_mutually_exclusive_group(required=True)
-    group_dat.add_argument('-d', '--dat-name', help='Select dat to update/check, must be in format "seed:name"')
-    group_dat.add_argument('-f', '--find', help='Select dats based on filter, they are "<field><operator><value>;...", valid operators are: =, !=, and ~=')
-    group_dat.add_argument('-a', '--all', help='Show all dats', action='store_true')
-
-    group_dat_action = parser_dat.add_mutually_exclusive_group(required=False)
-    group_dat_action.add_argument('-dt', '--details', help='Show details of dat', action='store_true')
-    group_dat_action.add_argument('-s', '--set', help='Manually set variable, must be in format "variable=value"')
-    group_dat_action.add_argument('--delete', action='store_true', default=False, help='Delete Dat')
-    group_dat_action.add_argument('--mark-mias', action='store_true', default=False, help='Mark Dat MIAs')
-
-    parser_dat.add_argument('-on', '--only-names', action='store_true', help='Only show names')
-
-    parser_dat.set_defaults(func=command_dat)
-
-    # Seed admin commands
-    parser_seed = subparser.add_parser('seed', help='Seed admin commands')
-    subparser_seed = parser_seed.add_subparsers(help='sub-command help')
-
-    parser_available = subparser_seed.add_parser('installed', help='List installed seeds')
-    parser_available.set_defaults(func=command_seed_installed)
-
-    parser_details = subparser_seed.add_parser('details', help='Show details of seed')
-    parser_details.add_argument('seed', help='Seed to show details of')
-    parser_details.set_defaults(func=command_seed_details)
-
-    parser_import = subparser.add_parser('import', help='Import dats from existing romvault')
-    parser_import.set_defaults(func=command_import)
-
-    parser_deduper = subparser.add_parser('deduper', help='Deduplicate dats, removes duplicates from input dat existing in parent dat')
-    parser_deduper.add_argument('-i', '--input', required=True, help='Input dat file e.g. "redump:psx_child" or "/mnt/roms/redump_psx_child.dat"')
-    parser_deduper.add_argument('-p', '--parent', default=None, help='Parent dat file e.g. "redump:psx" or "/mnt/roms/redump_psx.dat" if not set, parent is taken from input dat with prescanned dats')
-    parser_deduper.add_argument('-o', '--output', default=None, help='If different from input.dat, must be a file path e.g. "/mnt/roms/redump_psx_child_deduped.dat"')
-    parser_deduper.add_argument('-dr', '--dry-run', action='store_true', help='Do not write output file, just show actions')
-
-    parser_deduper.set_defaults(func=command_deduper)
-
-    def parse_seed(seed_name, description, seed=None):
-        parser_command = subparser.add_parser(seed_name, help=f'Seed {seed_name}, {description}')
-        parser_command.set_defaults(func=command_seed, seed=seed_name)
-        parser_command.add_argument('-d', '--details', action='store_true', help='Show details of seed')
-        parser_command.add_argument('-f', '--fetch', action='store_true', help='Fetch seed')
-        parser_command_process = parser_command.add_argument_group('process')
-        parser_command_process.add_argument('-p', '--process', action='store_true', help='Process dats from seed')
-        parser_command_process.add_argument('-a', '--actions', action='append', help='Action to execute')
-        parser_command_process.add_argument('-fd', '--filter', help='Filter dats to process')
-        if seed_name == 'all':
-            parser_command.add_argument('-e', '--exclude', action='append', help='Exclude seed or seeds (only work with all)')
-            parser_command.add_argument('-o', '--only', action='append', help='Only seed or seeds (only work with all)')
-        else:
-            seed.args(parser_command)
-
-    for seed in Seed.list_installed():
-        parse_seed(seed.name, seed.description(), seed=seed)
-    parse_seed('all', 'All seeds')
-
-
-    # Common arguments
-    subparsers = [subparser, subparser_seed]
-    for subpars in subparsers:
-        for subpar in subpars.choices.values():
-            subpar.add_argument('-v', '--verbose', action='store_true', help='verbose output')
-            subpar.add_argument('-q', '--quiet', action='store_true', help='quiet output')
-            subpar.add_argument('-nc', '--no-color', action='store_true', help='disable color output')
-            subpar.add_argument('-l', '--logging', action='store_true', help='enable logging')
+    add_log_parser(subparser)
+    add_config_parser(subparser)
+    add_doctor_parser(subparser)
+    add_dat_parser(subparser)
+    add_seed_parser(subparser)
+    add_import_parser(subparser)
+    add_deduper_parser(subparser)
+    add_all_seed_parser(subparser)
 
     if len(sys.argv) == 1:
         parser.print_help(sys.stderr)
@@ -139,7 +54,6 @@ def parse_args() -> argparse.Namespace:
     args = parser.parse_args()
     initial_setup(args)
     return args
-
 
 def initial_setup(args) -> None:
     """Initial setup of datoso from command line arguments."""
@@ -156,348 +70,6 @@ def initial_setup(args) -> None:
         config['COMMAND']['Verbose'] = 'true'
     if getattr(args, 'logging', False):
         enable_logging()
-
-
-def command_deduper(args) -> None:
-    """Deduplicate dats, removes duplicates from input dat existing in parent dat"""
-    if not args.parent and args.input.endswith(('.dat', '.xml')):
-        print('Parent dat is required when input is a dat file')
-        sys.exit(1)
-    if args.dry_run:
-        logger.setLevel(logging.DEBUG)
-    merged = Dedupe(args.input, args.parent)
-    merged.dedupe()
-    if args.output and not args.dry_run:
-        merged.save(args.output)
-    elif not args.dry_run:
-        merged.save()
-    logging.info('%s File saved to %s %s',
-                 Bcolors.OKBLUE,
-                 args.output if args.output else args.input,
-                 Bcolors.ENDC,
-                 )
-
-
-def command_import(_) -> None:
-    """Make changes in dat config"""
-    dat_root_path = config['PATHS']['DatPath']
-
-    if not dat_root_path or not Path(dat_root_path).exists():
-        print(f'{Bcolors.FAIL}Dat root path not set or does not exist{Bcolors.ENDC}')
-        sys.exit(1)
-
-    rules = Rules().rules
-
-    dats = { str(x):None for x in Path(dat_root_path).rglob('*.[dD][aA][tT]') }
-
-    if config['IMPORT'].get('IgnoreRegEx'):
-        ignore_regex = re.compile(config['IMPORT']['IgnoreRegEx'])
-        dats = [ dat for dat in dats if not ignore_regex.match(dat) ]
-
-    fromhere = ''
-    found = False
-    for dat_name in dats:
-        if fromhere in (dat_name, ''):
-            found = True
-        if not found:
-            continue
-        print(f'{dat_name} - ', end='')
-        seed, _class = detect_seed(dat_name, rules)
-        print(f'{seed} - {_class.__name__ if _class else None}')
-        if _class:
-            dat = _class(file=dat_name)
-            dat.load()
-            database = Dat(seed=seed, new_file=dat_name, **dat.dict())
-            database.save()
-            database.flush()
-
-
-def command_dat(args):
-    """Make changes in dat config"""
-
-    def print_dats(dats, fields = None):
-        """Print dats"""
-        output = []
-        fields = fields if fields else ['seed', 'name', 'status']
-        for dat in dats:
-            new_dat = { k:dat[k] for k in fields if dat.get(k) }
-            new_dat['status'] = dat.get('status', 'enabled') or 'enabled'
-            output.append(new_dat)
-        if getattr(args, 'only_names', False):
-            for dat in output:
-                print(f"{dat['seed']}:{dat['name']}")
-            return
-        print(tabulate(output, headers='keys', tablefmt='psql'))
-
-    from tinydb import Query
-
-    query = Query()
-    if args.dat_name:
-        splitted = args.dat_name.split(':')
-        if len(splitted) != 2:
-            print(f'{Bcolors.WARNING}Invalid dat name, must be in format "seed:name"{Bcolors.ENDC}')
-            print(f'Showing results for filter: {Bcolors.OKCYAN}name~={args.dat_name}{Bcolors.ENDC}')
-            print('--------------------------------------------------------------')
-            name = args.dat_name
-            result = Dat.search(query.name.matches(r'^.*' + name + r'.*', flags=re.IGNORECASE))
-            if args.details:
-                print_dats(result, fields=['name', 'modifier', 'company', 'system', 'seed', 'date', 'path', 'system_type', 'full_name', 'automerge', 'parent'])
-            else:
-                print_dats(result)
-        else:
-            seed, name = splitted
-            dat = Dat(seed=seed, name=name)
-            result = dat.get_one()
-            if not result:
-                print(f'{Bcolors.FAIL}Dat not found{Bcolors.ENDC}')
-                sys.exit(1)
-            if args.set:
-                key, value = args.set.split('=') if '=' in args.set else (args.set, True)
-                if value.isdigit():
-                    value = int(value)
-                if value.lower() == 'true':
-                    value = True
-                dat.update({key: value}, doc_ids=[result.doc_id])
-                dat.flush()
-                print(f'{Bcolors.OKGREEN}Dat {Bcolors.OKCYAN}{seed}:{name}{Bcolors.OKGREEN} {key} set to {Bcolors.OKBLUE}{value}{Bcolors.ENDC}')
-                sys.exit(0)
-            # ruff: noqa: ERA001
-            # TODO(laromicas): unset
-            # if args.unset:
-                #     key = args.unset
-                #     table.update({key: value}, doc_ids=[result.doc_id])
-                #     table.storage.flush()
-                #     print(f'{Bcolors.OKGREEN}Dat {Bcolors.OKCYAN}{seed}:{name}{Bcolors.OKGREEN} {key} set to {Bcolors.OKBLUE}{value}{Bcolors.ENDC}')
-                #     sys.exit(0)
-            if args.delete:
-                dat.remove(doc_ids=[result.doc_id])
-                dat.flush()
-                print(f'{Bcolors.OKGREEN}Dat {Bcolors.OKCYAN}{seed}:{name}{Bcolors.OKGREEN} removed{Bcolors.ENDC}')
-                sys.exit(0)
-            if args.mark_mias:
-                from datoso.mias.mia import mark_mias
-                mark_mias(dat_file=result['new_file'])
-                print(f'{Bcolors.OKGREEN}Set Dats {Bcolors.OKCYAN}{seed}:{name}{Bcolors.OKGREEN} marked MIAs{Bcolors.ENDC}')
-                sys.exit(0)
-            if args.details:
-                print_dats([result], fields=['name', 'modifier', 'company', 'system', 'seed', 'date', 'path', 'system_type', 'full_name', 'automerge', 'parent'])
-            else:
-                print_dats([result])
-    elif args.all:
-        # Show all dats
-        print_dats(dat.all())
-
-    elif args.find:
-        # Find dats TODO: finish
-        print(f'Showing results for filter: {Bcolors.OKCYAN}{args.find}{Bcolors.ENDC}')
-        print('--------------------------------------------------------------')
-        name, value = args.find.split('=')
-        from tinydb import where
-        result = Dat.search(where(name) == value)
-        print_dats(result)
-
-
-def command_seed_installed(_) -> None:
-    """List available seeds"""
-    print('Installed seeds:')
-    for seed, seed_module in installed_seeds().items():
-        description = seed_description(seed_module)
-        description = {description[0:60]+'...' if len(description) > 60 else description}
-        seed_name = seed[12:]
-        print(f'* {Bcolors.OKGREEN}{seed_name}{Bcolors.ENDC} - {description}')
-
-
-def command_seed_details(args) -> None:
-    module = None
-    for seed, seed_module in installed_seeds().items():
-        if seed == f'{__app_name__}_seed_{args.seed}':
-            module = seed_module
-            break
-    if not module:
-        print(f'Seed {Bcolors.FAIL}{args.seed}{Bcolors.ENDC} not installed')
-        sys.exit(1)
-    print(f'Seed {Bcolors.OKGREEN}{args.seed}{Bcolors.ENDC} details:')
-    print(f'  * Name: {module.__name__}')
-    print(f'  * Version: {module.__version__}')
-    print(f'  * Author: {module.__author__}')
-    print(f'  * Description: {module.__description__}')
-
-
-def command_seed(args) -> None:
-    """Commands with the seed (must be installed)"""
-    def parse_actions(args):
-        if args.actions and len(args.actions) == 1:
-                args.actions = args.actions[0].split(',')
-        if args.seed != 'all':
-            seed_object = Seed(name=args.seed)
-            if seed_object.parse_args:
-                seed_object.parse_args(args)
-        if not any([getattr(args, 'fetch', False), getattr(args, 'process', False), getattr(args, 'details', False)]):
-            print(f'{Bcolors.FAIL}No action specified{Bcolors.ENDC} (fetch, process, details)')
-            sys.exit(1)
-    parse_actions(args)
-    if args.seed == 'all':
-        for seed in installed_seeds():
-            seed_name = get_seed_name(seed)
-            if args.exclude and seed_name in args.exclude:
-                continue
-            if args.only and seed_name not in args.only:
-                continue
-            if config['PROCESS'].get('SeedIgnoreRegEx'):
-                ignore_regex = re.compile(config['PROCESS']['SeedIgnoreRegEx'])
-                if ignore_regex.match(seed_name):
-                    continue
-            args.seed = seed_name
-            command_seed(args)
-        sys.exit(0)
-    seed = Seed(name=args.seed)
-    if getattr(args, 'details', False):
-        command_seed_details(args)
-    if getattr(args, 'fetch', False):
-        message = f'{Bcolors.OKCYAN}Fetching seed {Bcolors.OKGREEN}{args.seed}{Bcolors.ENDC}'
-        print('='*(len(message)-14))
-        print(message)
-        print('='*(len(message)-14))
-        if seed.fetch():
-            print(f'Errors fetching {Bcolors.FAIL}{args.seed}{Bcolors.ENDC}')
-            print('Please enable logs for more information or use -v parameter')
-            command_doctor(args)
-            sys.exit(1)
-        print(f'{Bcolors.OKBLUE}Finished fetching {Bcolors.OKGREEN}{args.seed}{Bcolors.ENDC}')
-    if getattr(args, 'process', False):
-        message = f'{Bcolors.OKCYAN}Processing seed {Bcolors.OKGREEN}{args.seed}{Bcolors.ENDC}'
-        print('='*(len(message)-14))
-        print(message)
-        print('='*(len(message)-14))
-        if seed.process_dats(fltr=getattr(args, 'filter', None), actions_to_execute=args.actions):
-            print(f'Errors processing {Bcolors.FAIL}{args.seed}{Bcolors.ENDC}')
-            print('Please enable logs for more information or use -v parameter')
-            command_doctor(args)
-            sys.exit(1)
-        print(f'{Bcolors.OKBLUE}Finished processing {Bcolors.OKGREEN}{args.seed}{Bcolors.ENDC}')
-
-
-def command_config_save(args) -> None:
-    """Save config to file"""
-    config_file = Path('~/datosorc').expanduser() if args.directory == '~' else Path.cwd() / '.datosorc'
-    with open(config_file, 'w', encoding='utf-8') as file:
-        config.write(file)
-    print(f'Config saved to {Bcolors.OKGREEN}{config_file}{Bcolors.ENDC}')
-
-
-def command_config_set(args) -> None:
-    """Set config value, if global is set, it will be set in datoso.ini file"""
-    myconfig = args.set[0].split('.')
-    if len(myconfig) != 2:
-        print(f'{Bcolors.FAIL}Invalid config key, must be in <SECTION>.<Option> format. {Bcolors.ENDC}')
-        sys.exit(1)
-    if myconfig[1] not in config[myconfig[0]]:
-        print(f'{Bcolors.FAIL}Invalid config option. {Bcolors.ENDC}')
-        sys.exit(1)
-
-    newconfig = configparser.ConfigParser(comment_prefixes='/', allow_no_value=True)
-    newconfig.optionxform = lambda option: option
-    if getattr(args, 'global', False):
-        file = Path(ROOT_FOLDER) / 'datoso.ini'
-    else:
-        file = Path.cwd() / '.datosorc'
-        if not Path.is_file(file):
-            file = Path('~/datosorc').expanduser()
-    newconfig.read(file)
-    if not newconfig.has_section(myconfig[0]):
-        newconfig.add_section(myconfig[0])
-    newconfig[myconfig[0]][myconfig[1]] = args.set[1]
-    with open(file, 'w', encoding='utf-8') as file:
-        newconfig.write(file)
-    if getattr(args, 'global', False):
-        print(f'{Bcolors.OKGREEN}Global config {Bcolors.OKCYAN}{myconfig[0]}.{myconfig[1]}{Bcolors.OKGREEN} set to {Bcolors.OKBLUE}{args.set[1]}{Bcolors.ENDC}')
-    else:
-        print(f'{Bcolors.OKGREEN}Local Config {Bcolors.OKCYAN}{myconfig[0]}.{myconfig[1]}{Bcolors.OKGREEN} set to {Bcolors.OKBLUE}{args.set[1]}{Bcolors.ENDC}')
-
-
-def command_config_get(args) -> None:
-    """Get active config value"""
-    myconfig = args.get.split('.')
-    if len(myconfig) != 2:
-        print(myconfig)
-        print(f'{Bcolors.FAIL}Invalid config key, must be in <SECTION>.<Option> format. {Bcolors.ENDC}')
-        sys.exit(1)
-    if myconfig[1] not in config[myconfig[0]]:
-        print(f'{Bcolors.FAIL}Invalid config option. {Bcolors.ENDC}')
-        sys.exit(1)
-    print(config[myconfig[0]][myconfig[1]])
-
-
-def command_config_rules_update(args) -> None:
-    """Update rules from google sheet"""
-    from datoso.database.seeds import dat_rules
-    print('Updating rules')
-    try:
-        dat_rules.import_dats()
-        print('Rules updated')
-    except Exception as exc:
-        print(f'{Bcolors.FAIL}Error updating rules{Bcolors.ENDC}')
-        print(exc)
-        print('Please enable logs for more information or use -v parameter')
-        command_doctor(args)
-
-
-def command_config_mia_update(args) -> None:
-    """Update rules from google sheet"""
-    from datoso.database.seeds import mia
-    print('Updating MIA')
-    try:
-        mia.import_mias()
-        print('MIA updated')
-    except Exception as exc:
-        print(f'{Bcolors.FAIL}Error updating MIA{Bcolors.ENDC}')
-        print(exc)
-        print('Please enable logs for more information or use -v parameter')
-        command_doctor(args)
-
-
-def command_config(args) -> None:
-    """Config commands"""
-    if args.save:
-        command_config_save(args)
-    elif args.set:
-        command_config_set(args)
-    elif args.get:
-        command_config_get(args)
-    elif args.rules_update:
-        command_config_rules_update(args)
-    elif args.mia_update:
-        command_config_mia_update(args)
-    else:
-        config_dict = {s:dict(config.items(s)) for s in config.sections()}
-        print(json.dumps(config_dict, indent=4))
-
-
-def command_list(_):
-    """List installed seeds"""
-    for seed, seed_class in installed_seeds().items():
-        description = seed_class.description()
-        print(f'* {Bcolors.OKCYAN}{seed}{Bcolors.ENDC} - {description[0:60] if len(description) > 60 else description}...')
-
-def command_doctor(args):
-    """Doctor installed seeds"""
-    if getattr(args, 'seed', False):
-        seed = check_seed(args.seed)
-        if not seed:
-            print(f'Module Seed {Bcolors.FAIL}  - {Bcolors.BOLD}{args.seed}{Bcolors.ENDC} not found')
-            sys.exit(1)
-        seed = seed if args.seed.startswith('datoso_seed_') else f'datoso_seed_{args.seed}'
-        seeds = {seed: None}
-    else:
-        seeds = installed_seeds()
-    for seed, module in seeds.items():
-        print(f'Checking seed {Bcolors.OKCYAN}{seed}{Bcolors.ENDC}')
-        check_module(seed, module, args.repair)
-
-def command_log(_):
-    log_path = FileUtils.parse_folder(config.get('PATHS','DatosoPath', fallback='~/.datoso'))
-    logfile = log_path / config['LOG'].get('LogFile', 'datoso.log')
-    os.system(f'cat {logfile}')
 
 def main():
     """Main function"""
