+++ src/datoso/systems.json
@@ -236,6 +236,24 @@
         "system_type": "Handheld"
     },
     {
+        "company": null,
+        "system": "Bandai Little Jammer",
+        "override": {
+            "company": "Bandai",
+            "system": "Little Jammer"
+        },
+        "system_type": "Console"
+    },
+    {
+        "company": null,
+        "system": "Bandai Little Jammer Pro",
+        "override": {
+            "company": "Bandai",
+            "system": "Little Jammer Pro"
+        },
+        "system_type": "Console"
+    },
+    {
         "company": "Benesse",
         "system": "Pocket Challenge V2",
         "system_type": "Console"
